Moved Windows tests to a new project
