+++ .github/workflows/build.yml
@@ -18,6 +18,8 @@ jobs:
       run: dotnet build ./FO-DICOM.Core/FO-DICOM.Core.csproj --configuration Release --runtime win-x64
     - name: Test FO-DICOM.Tests
       run: dotnet test ./Tests/FO-DICOM.Tests/FO-DICOM.Tests.csproj --configuration Release --framework net462 --runtime win-x64 --logger:"trx;LogFileName=.\resultsnet462.xml"
+    - name: Test FO-DICOM.Tests
+      run: dotnet test ./Tests/FO-DICOM.Tests.Windows/FO-DICOM.Tests.Windows.csproj --configuration Release --framework net462 --runtime win-x64 --logger:"trx;LogFileName=.\resultsnet462.xml"
     - name: Upload test results
       uses: actions/upload-artifact@v4
       with:
@@ -30,6 +32,8 @@ jobs:
     - uses: actions/checkout@v4
     - name: Test FO-DICOM.Tests
       run: dotnet test ./Tests/FO-DICOM.Tests/FO-DICOM.Tests.csproj --configuration Release --framework net6.0-windows --blame --runtime win-x64 --logger:"trx;LogFileName=.\resultsnet60.xml"
+    - name: Test FO-DICOM.Tests
+      run: dotnet test ./Tests/FO-DICOM.Tests.Windows/FO-DICOM.Tests.Windows.csproj --configuration Release --framework net6.0-windows --blame --runtime win-x64 --logger:"trx;LogFileName=.\resultsnet60.xml"
     - name: Upload test results
       uses: actions/upload-artifact@v4
       with:
@@ -42,6 +46,8 @@ jobs:
     - uses: actions/checkout@v4
     - name: Test FO-DICOM.Tests
       run: dotnet test ./Tests/FO-DICOM.Tests/FO-DICOM.Tests.csproj --configuration Release --framework net8.0-windows --blame --runtime win-x64 --logger:"trx;LogFileName=.\resultsnet80.xml" --collect:"XPlat Code Coverage" --settings coverlet.runsettings
+    - name: Test FO-DICOM.Tests
+      run: dotnet test ./Tests/FO-DICOM.Tests.Windows/FO-DICOM.Tests.Windows.csproj --configuration Release --framework net8.0-windows --blame --runtime win-x64 --logger:"trx;LogFileName=.\resultsnet80.xml" --collect:"XPlat Code Coverage" --settings coverlet.runsettings
     - name: Upload test results
       uses: actions/upload-artifact@v4
       with:
