+++ FO-DICOM.Full.sln
@@ -9,6 +9,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{43ED1185
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FO-DICOM.Tests", "Tests\FO-DICOM.Tests\FO-DICOM.Tests.csproj", "{FB7608E8-FA90-4445-A7F2-430E6A518798}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FO-DICOM.Tests.Windows", "Tests\FO-DICOM.Tests.Windows\FO-DICOM.Tests.Windows.csproj", "{FB80E070-D958-4C1E-99AE-7304A110518D}"
+EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Serialization", "Serialization", "{3CAE7F87-9013-49B0-96B7-8C96055B8380}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FO-DICOM.Json", "Serialization\FO-DICOM.Json\FO-DICOM.Json.csproj", "{50FAD02E-6763-47A4-8592-A8BEC8658948}"
@@ -195,6 +197,18 @@ Global
 		{254100C7-5857-4760-94B7-5AEB8DBDDEED}.Release|x64.Build.0 = Release|Any CPU
 		{254100C7-5857-4760-94B7-5AEB8DBDDEED}.Release|x86.ActiveCfg = Release|Any CPU
 		{254100C7-5857-4760-94B7-5AEB8DBDDEED}.Release|x86.Build.0 = Release|Any CPU
+		{FB80E070-D958-4C1E-99AE-7304A110518D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FB80E070-D958-4C1E-99AE-7304A110518D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FB80E070-D958-4C1E-99AE-7304A110518D}.Debug|x64.ActiveCfg = Debug|x64
+		{FB80E070-D958-4C1E-99AE-7304A110518D}.Debug|x64.Build.0 = Debug|x64
+		{FB80E070-D958-4C1E-99AE-7304A110518D}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{FB80E070-D958-4C1E-99AE-7304A110518D}.Debug|x86.Build.0 = Debug|Any CPU
+		{FB80E070-D958-4C1E-99AE-7304A110518D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FB80E070-D958-4C1E-99AE-7304A110518D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FB80E070-D958-4C1E-99AE-7304A110518D}.Release|x64.ActiveCfg = Release|x64
+		{FB80E070-D958-4C1E-99AE-7304A110518D}.Release|x64.Build.0 = Release|x64
+		{FB80E070-D958-4C1E-99AE-7304A110518D}.Release|x86.ActiveCfg = Release|Any CPU
+		{FB80E070-D958-4C1E-99AE-7304A110518D}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -210,6 +224,7 @@ Global
 		{7C65BC75-0743-4364-9DAD-D9032BBBF3E1} = {A2EB9012-9955-4AAE-9AAC-C45A2D12BF92}
 		{1F89E5EB-302F-4779-8D3B-54B63E684697} = {368AEDCB-6CD6-471C-A62D-5C7F94DE5B4D}
 		{254100C7-5857-4760-94B7-5AEB8DBDDEED} = {CFA78DF7-BD2F-449E-8666-2FFE4358A789}
+		{FB80E070-D958-4C1E-99AE-7304A110518D} = {43ED1185-9895-4B31-8889-7C0477591BF1}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {4A7E4FCE-FCEB-485B-AE5F-ECF6918D5DA7}
