+++ Tests/FO-DICOM.Tests.Windows/WPFImageTest.cs
@@ -1,24 +1,20 @@
-﻿// Copyright (c) 2012-2023 fo-dicom contributors.
+﻿// Copyright (c) 2012-2024 fo-dicom contributors.
 // Licensed under the Microsoft Public License (MS-PL).
 #nullable disable
 
-#if WINDOWS
-using System.Windows.Media.Imaging;
-#endif
-
 using System.Drawing;
+using System.Windows.Media.Imaging;
 using FellowOakDicom.Imaging;
 using Xunit;
 
 namespace FellowOakDicom.Tests.Imaging
 {
 
-    [Collection(TestCollections.General)]
+    [Collection("Windows")]
     public class WPFImageTest
     {
 #region Unit tests
 
-#if WINDOWS
         [Fact]
         public void As_BitmapSource_ReturnsBitmapSource()
         {
@@ -34,8 +30,6 @@ namespace FellowOakDicom.Tests.Imaging
             image.Render(3, false, false, 0);
             Assert.Throws<DicomImagingException>(() => image.As<Bitmap>());
         }
-
-#endif
 
 #endregion
     }
