+++ Tests/FO-DICOM.Tests/IO/TemporaryFileTest.cs
@@ -99,19 +99,6 @@ namespace FellowOakDicom.Tests.IO
             }
         }
 
-#if WINDOWS
-        [Fact]
-        public void Create_StoragePathNull_FileAttributesContainTempFlag()
-        {
-            lock (_locker)
-            {
-                TemporaryFile.StoragePath = null;
-                var path = TemporaryFile.Create().Name;
-                Assert.True((File.GetAttributes(path) & FileAttributes.Temporary) == FileAttributes.Temporary);
-            }
-        }
-#endif
-
         #endregion
     }
 }
